fix(fiori-gen): move function calling template to src dir
