+++ packages/fiori-generator-shared/README.md
@@ -3,6 +3,7 @@
 Commonly used shared functionality to support the fiori generator.
 Add functionality that may be used by multiple generator modules.
 
+When adding additional templates, ensure the calling function (e.g., generateReadMe) remains at the src directory level to align with the template location pattern used during generator bundling.
 
 ## License
