+++ packages/fiori-generator-shared/src/index.ts
@@ -7,6 +7,6 @@ export * from './logging/logWrapper';
 export { getPackageScripts } from './npm-package-scripts/getPackageScripts';
 export { getBootstrapResourceUrls } from './ui5/ui5';
 export { getDefaultTargetFolder, isExtensionInstalled } from './vscode-helpers/vscode-helpers';
-export { generateReadMe } from './read-me/read-me';
+export { generateReadMe } from './read-me';
 export { getHostEnvironment } from './environment';
 export { YeomanEnvironment, VSCodeInstance, hostEnvironment, HostEnvironmentId } from './types';
