+++ packages/fiori-generator-shared/src/types/environment.ts
