+++ packages/fiori-generator-shared/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './read-me';
+export * from './environment';
