+++ packages/fiori-generator-shared/src/types/read-me.ts
